requirements.txt
celery==4.1.0
dj-database-url==0.4.2
Django==2.0.1
django-constance[database]==2.0.0
django-debug-toolbar==1.9.1
django-extensions==1.9.9
django-filter==1.1.0
django-guardian==1.4.9
django-rest-swagger==2.1.2
djangorestframework==3.7.7
Markdown==2.6.11
psycopg2==2.7.3.2
redis==2.10.6
pylibmc==1.5.2
